Move caption handling code to a usecase
